Streamlined consultation profile moves form and table to popups
Incorporated some of the feedback from Health Canada.  The web form is great but being in the sidebar makes it unclear how it relates to the page content.  I've moved the form into a pop-up overlay, and moved the table of meetings into another popup.  Keeps them both out of the way, making the page content more scannable.